+++ server/src/constants.ts
@@ -1,6 +1,153 @@
-
-export const HEXALOT_ALLOWED_RADII = {
-    7: true,
-}
-
 export const HEXALOT_PURCHASE_PRICE_SATOSHIS = "10"
+
+
+export const STOP_STEP = 0
+export const BRANCH_STEP = 7
+export const ERROR_STEP = 8
+
+
+export const HEXALOT_SHAPE = [
+    // center
+    {x: 0, y: 0},
+    // layer 1
+    {x: 2, y: 0}, // 1
+    {x: 1, y: -1},
+    {x: -1, y: -1},
+    {x: -2, y: 0},
+    {x: -1, y: 1},
+    {x: 1, y: 1}, // 6
+    // layer 2
+    {x: 4, y: 0}, // 7
+    {x: 3, y: -1},
+    {x: 2, y: -2},
+    {x: 0, y: -2},
+    {x: -2, y: -2},
+    {x: -3, y: -1},
+    {x: -4, y: 0},
+    {x: -3, y: 1},
+    {x: -2, y: 2},
+    {x: 0, y: 2},
+    {x: 2, y: 2},
+    {x: 3, y: 1}, // 18
+    // layer 3
+    {x: 6, y: 0}, // 19
+    {x: 5, y: -1},
+    {x: 4, y: -2},
+    {x: 3, y: -3},
+    {x: 1, y: -3},
+    {x: -1, y: -3},
+    {x: -3, y: -3},
+    {x: -4, y: -2},
+    {x: -5, y: -1},
+    {x: -6, y: 0},
+    {x: -5, y: 1},
+    {x: -4, y: 2},
+    {x: -3, y: 3},
+    {x: -1, y: 3},
+    {x: 1, y: 3},
+    {x: 3, y: 3},
+    {x: 4, y: 2},
+    {x: 5, y: 1}, // 36
+    // layer 4
+    {x: 8, y: 0}, // 37
+    {x: 7, y: -1},
+    {x: 6, y: -2},
+    {x: 5, y: -3},
+    {x: 4, y: -4},
+    {x: 2, y: -4},
+    {x: 0, y: -4},
+    {x: -2, y: -4},
+    {x: -4, y: -4},
+    {x: -5, y: -3},
+    {x: -6, y: -2},
+    {x: -7, y: -1},
+    {x: -8, y: 0},
+    {x: -7, y: 1},
+    {x: -6, y: 2},
+    {x: -5, y: 3},
+    {x: -4, y: 4},
+    {x: -2, y: 4},
+    {x: -0, y: 4},
+    {x: 2, y: 4},
+    {x: 4, y: 4},
+    {x: 5, y: 3},
+    {x: 6, y: 2},
+    {x: 7, y: 1}, // 60
+    // layer 5
+    {x: 10, y: 0}, // 61
+    {x: 9, y: -1},
+    {x: 8, y: -2},
+    {x: 7, y: -3},
+    {x: 6, y: -4},
+    {x: 5, y: -5},
+    {x: 3, y: -5},
+    {x: 1, y: -5},
+    {x: -1, y: -5},
+    {x: -3, y: -5},
+    {x: -5, y: -5},
+    {x: -6, y: -4},
+    {x: -7, y: -3},
+    {x: -8, y: -2},
+    {x: -9, y: -1},
+    {x: -10, y: 0},
+    {x: -9, y: 1},
+    {x: -8, y: 2},
+    {x: -7, y: 3},
+    {x: -6, y: 4},
+    {x: -5, y: 5},
+    {x: -3, y: 5},
+    {x: -1, y: 5},
+    {x: 1, y: 5},
+    {x: 3, y: 5},
+    {x: 5, y: 5},
+    {x: 6, y: 4},
+    {x: 7, y: 3},
+    {x: 8, y: 2},
+    {x: 9, y: 1}, // 90
+    // layer 6
+    {x: 12, y: 0}, // 91
+    {x: 11, y: -1},
+    {x: 10, y: -2},
+    {x: 9, y: -3},
+    {x: 8, y: -4},
+    {x: 7, y: -5},
+    {x: 6, y: -6},
+    {x: 4, y: -6},
+    {x: 2, y: -6},
+    {x: 0, y: -6},
+    {x: -2, y: -6},
+    {x: -4, y: -6},
+    {x: -6, y: -6},
+    {x: -7, y: -5},
+    {x: -8, y: -4},
+    {x: -9, y: -3},
+    {x: -10, y: -2},
+    {x: -11, y: -1},
+    {x: -12, y: 0},
+    {x: -11, y: 1},
+    {x: -10, y: 2},
+    {x: -9, y: 3},
+    {x: -8, y: 4},
+    {x: -7, y: 5},
+    {x: -6, y: 6},
+    {x: -4, y: 6},
+    {x: -2, y: 6},
+    {x: 0, y: 6},
+    {x: 2, y: 6},
+    {x: 4, y: 6},
+    {x: 6, y: 6},
+    {x: 7, y: 5},
+    {x: 8, y: 4},
+    {x: 9, y: 3},
+    {x: 10, y: 2},
+    {x: 11, y: 1}, // 126
+]
+
+export const ADJACENT = [
+    {x: 2, y: 0}, // 1
+    {x: 1, y: -1},
+    {x: -1, y: -1},
+    {x: -2, y: 0},
+    {x: -1, y: 1},
+    {x: 1, y: 1}, // 6
+]
