+++ server/src/types.ts
@@ -1,3 +1,2 @@
 
 export type HexalotID = string
-export type PubKey = string
